dedupe some code
