+++ src/utils/recurrance.ts
@@ -1,0 +1,5 @@
+export type IntervalUnit = 'hours' | 'days' | 'weeks' | 'months' | 'years'
+
+export const INTERVAL_UNITS: IntervalUnit[] = ['hours', 'days', 'weeks', 'months', 'years']
+
+export type FrequencyType = 'once' | 'interval' | 'days_of_the_week' | 'day_of_the_month' | 'custom'
