+++ src/views/Chores/ChoreCard.tsx
@@ -22,7 +22,6 @@ import {
   MenuItem,
   Typography,
 } from '@mui/joy'
-import moment from 'moment'
 import React from 'react'
 import { getTextColorFromBackgroundColor } from '../../utils/Colors'
 import {
@@ -35,7 +34,7 @@ import { ConfirmationModal, ConfirmationModalProps } from '../Modals/Inputs/Conf
 import { DateModal } from '../Modals/Inputs/DateModal'
 import { SxProps } from '@mui/joy/styles/types'
 import { withNavigation } from '../../contexts/hooks'
-import { Chore } from '../../models/chore'
+import { Chore, getDueDateChipColor, getDueDateChipText, getRecurrentChipText } from '../../models/chore'
 
 interface ChoreCardProps {
   chore: Chore
@@ -131,134 +130,14 @@ class ChoreCardInner extends React.Component<ChoreCardProps, ChoreCardState> {
     )
   }
 
-  private getDueDateChipText = nextDueDate => {
-    if (this.props.chore.nextDueDate === null) return 'No Due Date'
-    // if due in next 48 hours, we should it in this format : Tomorrow 11:00 AM
-    const diff = moment(nextDueDate).diff(moment(), 'hours')
-    if (diff < 48 && diff > 0) {
-      return moment(nextDueDate).calendar().replace(' at', '')
-    }
-    return 'Due ' + moment(nextDueDate).fromNow()
-  }
-
-  private getDueDateChipColor = nextDueDate => {
-    if (this.props.chore.nextDueDate === null) return 'neutral'
-    const diff = moment(nextDueDate).diff(moment(), 'hours')
-    if (diff < 48 && diff > 0) {
-      return 'warning'
-    }
-    if (diff < 0) {
-      return 'danger'
-    }
-
-    return 'neutral'
-  }
-
-  private getRecurrentChipText = chore => {
-    const dayOfMonthSuffix = n => {
-      if (n >= 11 && n <= 13) {
-        return 'th'
-      }
-      switch (n % 10) {
-        case 1:
-          return 'st'
-        case 2:
-          return 'nd'
-        case 3:
-          return 'rd'
-        default:
-          return 'th'
-      }
-    }
-    if (chore.frequencyType === 'once') {
-      return 'Once'
-    } else if (chore.frequencyType === 'trigger') {
-      return 'Trigger'
-    } else if (chore.frequencyType === 'daily') {
-      return 'Daily'
-    } else if (chore.frequencyType === 'adaptive') {
-      return 'Adaptive'
-    } else if (chore.frequencyType === 'weekly') {
-      return 'Weekly'
-    } else if (chore.frequencyType === 'monthly') {
-      return 'Monthly'
-    } else if (chore.frequencyType === 'yearly') {
-      return 'Yearly'
-    } else if (chore.frequencyType === 'days_of_the_week') {
-      let days = JSON.parse(chore.frequencyMetadata).days
-      if (days.length > 4) {
-        const allDays = [
-          'Sunday',
-          'Monday',
-          'Tuesday',
-          'Wednesday',
-          'Thursday',
-          'Friday',
-          'Saturday',
-        ]
-        const selectedDays = days.map(d => moment().day(d).format('dddd'))
-        const notSelectedDay = allDays.filter(
-          day => !selectedDays.includes(day),
-        )
-        const notSelectedShortdays = notSelectedDay.map(d =>
-          moment().day(d).format('ddd'),
-        )
-        return `Daily except ${notSelectedShortdays.join(', ')}`
-      } else {
-        days = days.map(d => moment().day(d).format('ddd'))
-        return days.join(', ')
-      }
-    } else if (chore.frequencyType === 'day_of_the_month') {
-      const months = JSON.parse(chore.frequencyMetadata).months
-      if (months.length > 6) {
-        const allMonths = [
-          'January',
-          'February',
-          'March',
-          'April',
-          'May',
-          'June',
-          'July',
-          'August',
-          'September',
-          'October',
-          'November',
-          'December',
-        ]
-        const selectedMonths = months.map(m => moment().month(m).format('MMMM'))
-        const notSelectedMonth = allMonths.filter(
-          month => !selectedMonths.includes(month),
-        )
-        const notSelectedShortMonths = notSelectedMonth.map(m =>
-          moment().month(m).format('MMM'),
-        )
-        return `${chore.frequency}${dayOfMonthSuffix(
-          chore.frequency,
-        )} except ${notSelectedShortMonths.join(', ')}`
-      } else {
-        const freqData = JSON.parse(chore.frequencyMetadata)
-        const months = freqData.months.map(m => moment().month(m).format('MMM'))
-        return `${chore.frequency}${dayOfMonthSuffix(
-          chore.frequency,
-        )} of ${months.join(', ')}`
-      }
-    } else if (chore.frequencyType === 'interval') {
-      return `Every ${chore.frequency} ${
-        JSON.parse(chore.frequencyMetadata).unit
-      }`
-    } else {
-      return chore.frequencyType
-    }
-  }
-
   private getFrequencyIcon = chore => {
-    if (['once', 'no_repeat'].includes(chore.frequencyType)) {
-      return <TimesOneMobiledata />
-    } else if (chore.frequencyType === 'trigger') {
-      return <Webhook />
-    } else {
-      return <Repeat />
-    }
+      if (['once', 'no_repeat'].includes(chore.frequencyType)) {
+          return <TimesOneMobiledata />
+      } else if (chore.frequencyType === 'trigger') {
+          return <Webhook />
+      } else {
+          return <Repeat />
+      }
   }
 
   private getName = name => {
@@ -288,9 +167,9 @@ class ChoreCardInner extends React.Component<ChoreCardProps, ChoreCardState> {
             zIndex: 1,
             left: 10,
           }}
-          color={this.getDueDateChipColor(chore.nextDueDate)}
+          color={getDueDateChipColor(chore.nextDueDate)}
         >
-          {this.getDueDateChipText(chore.nextDueDate)}
+          {getDueDateChipText(chore.nextDueDate)}
         </Chip>
 
         <Chip
@@ -311,7 +190,7 @@ class ChoreCardInner extends React.Component<ChoreCardProps, ChoreCardState> {
             }}
           >
             {this.getFrequencyIcon(chore)}
-            {this.getRecurrentChipText(chore)}
+            {getRecurrentChipText(chore)}
           </div>
         </Chip>
